requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
Pillow==10.1.0
python-decouple==3.8
django-filter==23.5
celery==5.3.4
redis==5.0.1
stripe==7.8.0

# Production dependencies for Railway hosting
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
